+++ notebooks/py_scripts/04-Automate_Optional.py
@@ -30,14 +30,18 @@ if os.path.basename(os.getcwd())=='notebooks': os.chdir('..')
 assert os.path.basename(os.getcwd())=='workflow-deeplabcut', ("Please move to the "
                                                               + "workflow directory")
 from workflow_deeplabcut.pipeline import lab, subject, session, train, model
+from workflow_deeplabcut import process
 
 # %% [markdown]
+# We'll be using the `process.py` script to automatically loop through all `make` functions, as a shortcut for calling each individually.
+#
 # If you previously completed the [03-Process notebook](./03-Process.ipynb), you may want to delete the contents ingested there, to avoid duplication errors.
 
 # %%
-# (session.Session & 'subject="subject6"').delete()
-# train.TrainingParamSet.delete()
-# train.VideoSet.delete()
+safemode=None # Set to false to turn off confirmation prompts
+(session.Session & 'subject="subject6"').delete(safemode=safemode)
+train.TrainingParamSet.delete(safemode=safemode)
+train.VideoSet.delete(safemode=safemode)
 
 # %% [markdown]
 # ## Ingestion of subjects, sessions, videos and training parameters
@@ -55,7 +59,6 @@ from workflow_deeplabcut.pipeline import lab, subject, session, train, model
 #         - `train.ModelTrainingParamSet`
 #         - `train.VideoSet` and the corresponding `File` part table
 #         - `model.VideoRecording` and the corresponding `File` part table
-#         - `model.RecordingInfo`
 
 # %%
 from workflow_deeplabcut.ingest import ingest_subjects, ingest_sessions, ingest_dlc_items
@@ -74,53 +77,52 @@ data_dir = find_full_path(dj.config['custom']['dlc_root_data_dir'], # root from 
 config_path = (data_dir / 'config.yaml')
 
 # %% [markdown]
-# 2. For the purposes of this demo, we'll ask DeepLabCut to structure the demo config file with `load_demo_data`. If you already did this in the [00-DataDownload notebook](./00-DataDownload_Optional.ipynb), skip this step.
+# 2. For the purposes of this demo, we will
+#    1. ask DeepLabCut to structure the demo config file with `load_demo_data`. If you already did this in the [00-DataDownload notebook](./00-DataDownload_Optional.ipynb), skip this step.
+#    2. generate a copy to show pose estimation. This is `recording_id` 2 in `recordings.csv`. If you already did this in the [00-DataDownload notebook](./00-DataDownload_Optional.ipynb), skip this step.
 
 # %%
+# A
 from deeplabcut.create_project.demo_data import load_demo_data
-load_demo_data(config_path)
-
-# %% [markdown]
-# 3. For this demo, we generate a copy to show pose estimation. This is `recording_id` 2 in `recordings.csv`. If you already did this in the [00-DataDownload notebook](./00-DataDownload_Optional.ipynb), skip this step.
-
-# %% tags=[]
+# load_demo_data(config_path)
+# B
 vid_path = str(data_dir).replace(" ", "\ ") + '/videos/m3v1mp4'
 cmd = (f'ffmpeg -n -hide_banner -loglevel error -ss 0 -t 2 -i {vid_path}.mp4 -vcodec copy '
        + f'-acodec copy {vid_path}-copy.mp4') # New video copy, first 2 seconds
 os.system(cmd)
 
 # %% [markdown]
-# 4. Pair training files with training parameters, and launch training.
+# 3. Next, we pair training files with training parameters, and launch training via `process`. 
+#    - Some tables may try to populate without upstream keys. 
+#    - Others, like `RecordingInfo` already have keys loaded.
+#    - Note: DLC's model processes (e.g., Training, Evaluation) log a lot of information to the console, to quiet this, pass `verbose=False` to `process`
 
 # %%
 key={'paramset_idx':1,'training_id':1,'video_set_id':1, 
      'project_path':'openfield-Pranav-2018-10-30/'}
 train.TrainingTask.insert1(key, skip_duplicates=True)
-train.ModelTraining.populate()
+process.run(verbose=True)
+model.RecordingInfo()
 
 # %% [markdown]
-# 5. Add this model to the `Model` table and evaluate.
+# 4. Now to add such an upstream key: a model to the `Model` table, and `process` to evaluate.
 #    - Include a user-friendly `model_name`
 #    - Include the relative path for the project's `config.yaml`
 #    - Add `shuffle` and `trainingsetindex`
 #    - `insert_new_model` will prompt before inserting, but this can be skipped with `prompt=False`
 
 # %%
-model.Model.insert_new_model(model_name='OpenField-5',dlc_config=config_path,
-                             shuffle=1,trainingsetindex=0, paramset_idx=1, 
+model.Model.insert_new_model(model_name='OpenField-5', 
+                             dlc_config=config_path,
+                             shuffle=1,
+                             trainingsetindex=0,
+                             paramset_idx=1, 
                              prompt=True, # True is the default behavior
                              model_description='Open field model trained 5 iterations')
-model.ModelEvaluation.populate()
+process.run()
 
 # %% [markdown]
-# 6. Auto-populate the `RecordingInfo` table with relevant fields.
-
-# %%
-model.RecordingInfo.populate()
-model.RecordingInfo()
-
-# %% [markdown]
-# 7. Add a pose estimation task, and launch pose estimation.
+# 5. Add a pose estimation task, and launch via `process`.
 #    - Get all primary key information for a given recording
 #    - Add the model and `task_mode` (i.e., load vs. trigger) to be applied
 #    - Add any additional analysis parameters for `deeplabcut.analyze_videos`
@@ -130,10 +132,10 @@ key=(model.VideoRecording & 'recording_id=2').fetch1('KEY')
 key.update({'model_name': 'OpenField-5', 'task_mode': 'trigger'})
 analyze_params={'save_as_csv':True} # add any others from deeplabcut.analyze_videos
 model.PoseEstimationTask.insert_estimation_task(key,params=analyze_params)
-model.PoseEstimation.populate()
+process.run()
 
 # %% [markdown]
-# 8. Retrieve estimated position data.
+# 6. Retrieve estimated position data.
 
 # %%
 model.PoseEstimation.get_trajectory(key)