+++ workflow_deeplabcut/ingest.py
@@ -1,31 +1,42 @@
 # from pathlib import Path
 import csv
 import ruamel.yaml as yaml
-from element_interface.utils import find_full_path
+from element_interface.utils import find_full_path  # , ingest_csv_to_table
 
 from .pipeline import subject, session, train, model
 from .paths import get_dlc_root_data_dir
 
 
-def ingest_general(csvs, tables,
-                   skip_duplicates=True):
+## TODO: why did pip install git+URL not have this function from element-interface?
+def ingest_csv_to_table(csvs, tables, verbose=True, skip_duplicates=True,
+                        ignore_extra_fields=True, allow_direct_insert=False):
     """
     Inserts data from a series of csvs into their corresponding table:
-        e.g., ingest_general(['./lab.csv', './subject.csv'],
-                                 [lab.Lab(),subject.Subject()]
-    ingest_general(csvs, tables, skip_duplicates=True)
-        :param csvs: list of relative paths to CSV files
+        e.g., ingest_csv_to_table(['./lab_data.csv', './proj_data.csv'],
+                                 [lab.Lab(),lab.Project()]
+    ingest_csv_to_table(csvs, tables, skip_duplicates=True)
+        :param csvs: list of relative paths to CSV files.  CSV are delimited by commas.
         :param tables: list of datajoint tables with ()
+        :param verbose: print number inserted (i.e., table length change)
+        :param skip_duplicates: skip duplicate entries
+        :param ignore_extra_fields: if a csv feeds multiple tables, the subset of
+                                    columns not applicable to a table will be ignored
+        :param allow_direct_insert: permit insertion into Imported and Computed tables
     """
-    for insert, table in zip(csvs, tables):
-        with open(insert, newline='') as f:
+    for csv_filepath, table in zip(csvs, tables):
+        with open(csv_filepath, newline='') as f:
             data = list(csv.DictReader(f, delimiter=','))
-        prev_len = len(table)
+        if verbose:
+            prev_len = len(table)
         table.insert(data, skip_duplicates=skip_duplicates,
-                     ignore_extra_fields=True)
-        insert_len = len(table) - prev_len     # report length change
-        print(f'\n---- Inserting {insert_len} entry(s) '
-              + f'into {table.table_name} ----')
+                     # Ignore extra fields because some CSVs feed multiple tables
+                     ignore_extra_fields=ignore_extra_fields,
+                     # Allow direct bc element-event uses dj.Imported w/o `make` funcs
+                     allow_direct_insert=allow_direct_insert)
+        if verbose:
+            insert_len = len(table) - prev_len
+            print(f'\n---- Inserting {insert_len} entry(s) '
+                  + f'into {table.table_name} ----')
 
 
 def ingest_subjects(subject_csv_path='./user_data/subjects.csv',
@@ -38,7 +49,7 @@ def ingest_subjects(subject_csv_path='./user_data/subjects.csv',
     """
     csvs = [subject_csv_path]
     tables = [subject.Subject()]
-    ingest_general(csvs, tables, skip_duplicates=skip_duplicates)
+    ingest_csv_to_table(csvs, tables, skip_duplicates=skip_duplicates)
 
 
 def ingest_sessions(session_csv_path='./user_data/sessions.csv',
@@ -50,7 +61,7 @@ def ingest_sessions(session_csv_path='./user_data/sessions.csv',
     tables = [session.Session(), session.SessionDirectory(),
               session.SessionNote()]
 
-    ingest_general(csvs, tables, skip_duplicates=skip_duplicates)
+    ingest_csv_to_table(csvs, tables, skip_duplicates=skip_duplicates)
 
 
 def ingest_dlc_items(config_params_csv_path='./user_data/config_params.csv',
@@ -93,14 +104,7 @@ def ingest_dlc_items(config_params_csv_path='./user_data/config_params.csv',
             model_video_csv_path, model_video_csv_path]
     tables = [train.VideoSet(), train.VideoSet.File(),
               model.VideoRecording(), model.VideoRecording.File()]
-    ingest_general(csvs, tables, skip_duplicates=skip_duplicates)
-
-    # Populate RecordingInfo
-    previous_length = len(model.RecordingInfo.fetch())
-    model.RecordingInfo.populate()
-    insert_length = len(model.RecordingInfo.fetch()) - previous_length
-    print(f'\n---- Inserting {insert_length} entry(s) into _recording_info ----')
-
+    ingest_csv_to_table(csvs, tables, skip_duplicates=skip_duplicates)
 
 if __name__ == '__main__':
     ingest_subjects()